# apps/api/openapi.spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/projects/{id}": {
      "get": {
        "operationId": "ProjectsController_findOne",
        "summary": "Find a project with id",
        "description": "Find a project with id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The id of the project",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericErrorEntity"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericErrorEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Projects"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
    "/projects/{id}/logs": {
      "post": {
        "operationId": "LogsController_create",
        "summary": "Create Logs",
        "description": "Create logs",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Id of project",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLogDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericErrorEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Logs"
        ],
        "security": [
          {
            "Api-Key": []
          }
        ]
      }
    },
    "/projects/{id}/request-logs": {
      "get": {
        "operationId": "LogsController_getRequestLogs",
        "summary": "Get requests logs for a project",
        "description": "Get requests logs for a project",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Id of project",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "The no of page",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "The no of items to fetch, defaults to 6",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "The search query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "apiKey",
            "required": false,
            "in": "query",
            "description": "The api keys",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "fromDate",
            "required": false,
            "in": "query",
            "description": "The from date",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "toDate",
            "required": false,
            "in": "query",
            "description": "The to date",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "method",
            "required": false,
            "in": "query",
            "description": "The method",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "statusCode",
            "required": false,
            "in": "query",
            "description": "The status code",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FetchRequestLogsResponseEntity"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericErrorEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Logs"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/projects/{id}/request-logs/{logId}": {
      "get": {
        "operationId": "LogsController_getRequestLog",
        "summary": "Get a request log by id",
        "description": "Get a request log by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Id of request log",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "logId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogEntity"
                }
              }
            }
          },
          "404": {
            "description": "Log or Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericErrorEntity"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericErrorEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Logs"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/projects/{id}/application-logs": {
      "get": {
        "operationId": "LogsController_getApplicationLogs",
        "summary": "Get application logs for a project",
        "description": "Get application logs for a project",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Id of project",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "The no of page",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "The no of items to fetch, defaults to 6",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "The search query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fromDate",
            "required": false,
            "in": "query",
            "description": "The from date",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "toDate",
            "required": false,
            "in": "query",
            "description": "The to date",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "level",
            "required": false,
            "in": "query",
            "description": "The level",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "apiKey",
            "required": false,
            "in": "query",
            "description": "The api keys",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "component",
            "required": false,
            "in": "query",
            "description": "The component",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "functionName",
            "required": false,
            "in": "query",
            "description": "The function name",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FetchApplicationLogsResponseEntity"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericErrorEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Logs"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/projects/{id}/application-logs/{logId}": {
      "get": {
        "operationId": "LogsController_getApplicationLog",
        "summary": "Get an application log by id",
        "description": "Get an application log by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Id of project",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "logId",
            "required": true,
            "in": "path",
            "description": "Id of log",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtendedApplicationLogEntity"
                }
              }
            }
          },
          "404": {
            "description": "Log or Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericErrorEntity"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericErrorEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Logs"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "LogsTrap",
    "description": "A functional logging solution",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header",
        "x-tokenName": "Token"
      },
      "Api-Key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "GenericErrorEntity": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "ProjectMode": {
        "type": "string",
        "enum": [
          "test",
          "live"
        ]
      },
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "mode": {
            "$ref": "#/components/schemas/ProjectMode"
          }
        },
        "required": [
          "id",
          "name",
          "mode"
        ]
      },
...
      "PartialApiKeyEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
...
      "CreateApplicationLogDto": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "description": "Timestamp of the log entry",
            "format": "date-time"
          },
          "level": {
            "$ref": "#/components/schemas/LogLevel"
          },
          "message": {
            "type": "string",
            "description": "Log message"
          },
          "component": {
            "type": "string",
            "description": "Component that generated the log"
          },
          "functionName": {
            "type": "string",
            "description": "Function name where the log was generated"
          },
          "additionalInfo": {
            "type": "object",
            "description": "Additional information"
          }
        },
        "required": [
          "level",
          "message"
        ]
      },
      "RequestLogDTO": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "date-time",
            "description": "timestamp when log was created"
          },
          "timeTaken": {
            "type": "number",
            "description": "Time taken for your service to send response"
          },
          "method": {
            "type": "string",
            "description": "Request method"
          },
          "url": {
            "type": "string",
            "description": "The url where request was made(/api/v1..., https://api.com/v1/)"
          },
          "statusCode": {
            "type": "number",
            "description": "Status code of this request"
          },
          "requestBody": {
            "type": "object",
            "description": "The request body(only supports json)"
          },
          "responseBody": {
            "type": "object",
            "description": "The response body(only supports json)"
          },
          "requestHeaders": {
            "type": "object",
            "description": "The request headers(only supports json)"
          },
          "responseHeaders": {
            "type": "object",
            "description": "The response headers(only supports json)"
          },
          "cookies": {
            "type": "object",
            "description": "The request cookies(only supports json)"
          },
          "ip": {
            "type": "string",
            "description": "The ip from where the request was made"
          },
          "userAgent": {
            "type": "string",
            "description": "The user agent from where the request was made"
          },
          "host": {
            "type": "string",
            "description": "The host from where the request was made"
          },
          "applicationLogs": {
            "description": "Application Logs related to this request",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateApplicationLogDto"
            }
          }
        },
        "required": [
          "method",
          "url",
          "statusCode"
        ]
      },
      "CreateSystemLogDto": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "description": "Timestamp of the log entry",
            "format": "date-time"
          },
          "level": {
            "$ref": "#/components/schemas/LogLevel"
          },
          "message": {
            "type": "string",
            "description": "Log message"
          },
          "eventType": {
            "type": "string",
            "description": "Type of system event"
          },
          "details": {
            "type": "object",
            "description": "Additional details"
          }
        },
        "required": [
          "level",
          "message"
        ]
      },
      "CreateLogDto": {
        "type": "object",
        "properties": {
          "requests": {
            "description": "Requests associated to this log",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RequestLogDTO"
            }
          },
          "applicationLogs": {
            "description": "Application logs",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateApplicationLogDto"
            }
          },
          "systemLogs": {
            "description": "System logs(like when a process crashes)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateSystemLogDto"
            }
          }
        }
      },
      "PartialRequestLogEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "timestamp": {
            "type": "string"
          },
          "userAgent": {
            "type": "string"
          },
          "applicationLogsCount": {
            "type": "number"
          },
          "method": {
            "type": "string"
          },
          "statusCode": {
            "type": "number"
          },
          "apiKeyId": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "timeTaken": {
            "type": "string"
          },
          "apiKeyName": {
            "type": "string"
          },
          "host": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "url",
          "timestamp",
          "userAgent",
          "applicationLogsCount",
          "method",
          "statusCode",
          "apiKeyId",
          "projectId"
        ]
      },
      "FetchRequestLogsResponseEntity": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PartialRequestLogEntity"
            }
          },
          "totalItems": {
            "type": "number"
          },
          "itemsPerQuery": {
            "type": "number",
            "default": 6
          }
        },
        "required": [
          "items",
          "totalItems",
          "itemsPerQuery"
        ]
      },
      "ApplicationLogEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "level": {
            "type": "string",
            "enum": [
              "debug",
              "info",
              "warn",
              "error",
              "fatal",
              "log",
              "trace"
            ]
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          },
          "additionalInfo": {
            "type": "string",
            "description": "Additional info(encrypted)"
          },
          "functionName": {
            "type": "string"
          },
          "component": {
            "type": "string"
          },
          "iv": {
            "type": "string",
            "description": "Initialization vector"
          }
        },
        "required": [
          "id",
          "level",
          "message",
          "timestamp",
          "iv"
        ]
      },
      "LogEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "timestamp": {
            "type": "string"
          },
          "userAgent": {
            "type": "string"
          },
          "applicationLogsCount": {
            "type": "number"
          },
          "method": {
            "type": "string"
          },
          "statusCode": {
            "type": "number"
          },
          "apiKeyId": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "timeTaken": {
            "type": "string"
          },
          "apiKeyName": {
            "type": "string"
          },
          "host": {
            "type": "string"
          },
          "requestBody": {
            "type": "string",
            "description": "Request body(encrypted)"
          },
          "responseBody": {
            "type": "string",
            "description": "Response body(encrypted)"
          },
          "requestHeaders": {
            "type": "string",
            "description": "Request headers(encrypted)"
          },
          "responseHeaders": {
            "type": "string",
            "description": "Response headers(encrypted)"
          },
          "cookies": {
            "type": "string",
            "description": "Cookies(encrypted)"
          },
          "ip": {
            "type": "string",
            "description": "IP address"
          },
          "applicationLogs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApplicationLogEntity"
            }
          },
          "iv": {
            "type": "string",
            "description": "Initialization vector"
          }
        },
        "required": [
          "id",
          "url",
          "timestamp",
          "userAgent",
          "applicationLogsCount",
          "method",
          "statusCode",
          "projectId",
          "requestBody",
          "iv"
        ]
      },
...
      "FetchApplicationLogsResponseEntity": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PartialApplicationLogEntity"
            }
          },
          "totalItems": {
            "type": "number"
          },
          "itemsPerQuery": {
            "type": "number",
            "default": 6
          }
        },
        "required": [
          "items",
          "totalItems",
          "itemsPerQuery"
        ]
      },
      "LogEntityWithIdAndUrl": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "url"
        ]
      },
      "ExtendedApplicationLogEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "level": {
            "type": "string",
            "enum": [
              "debug",
              "info",
              "warn",
              "error",
              "fatal",
              "log",
              "trace"
            ]
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          },
          "additionalInfo": {
            "type": "string",
            "description": "Additional info(encrypted)"
          },
          "functionName": {
            "type": "string"
          },
          "component": {
            "type": "string"
          },
          "iv": {
            "type": "string",
            "description": "Initialization vector"
          },
          "apiKey": {
            "$ref": "#/components/schemas/PartialApiKeyEntity"
          },
          "request": {
            "$ref": "#/components/schemas/LogEntityWithIdAndUrl"
          }
        },
        "required": [
          "id",
          "level",
          "message",
          "timestamp",
          "iv",
          "apiKey",
          "request"
        ]
      }
    }
  }
}